+++ packages/cli/src/commands/ui/create/react.spec.ts
@@ -17,7 +17,7 @@ import {spawnProcessOutput} from '../../../lib/utils/process';
 import {npxInPty} from '../../../lib/utils/npx';
 import {AuthenticatedClient} from '../../../lib/platform/authenticatedClient';
 import PlatformClient from '@coveord/platform-client';
-import {Config, Configuration} from '../../../lib/config/config';
+import {Config} from '../../../lib/config/config';
 import {
   IsNpxInstalled,
   IsNodeVersionInRange,
@@ -27,6 +27,7 @@ import {getPackageVersion} from '../../../lib/utils/misc';
 import Command from '@oclif/command';
 import {appendCmdIfWindows} from '../../../lib/utils/os';
 import {IPty} from 'node-pty';
+import {configurationMock} from '../../../__stub__/configuration';
 
 describe('ui:create:react', () => {
   const mockedConfig = mocked(Config);
@@ -71,7 +72,7 @@ describe('ui:create:react', () => {
     mockedSpawnProcessOutput.mockResolvedValue({
       stdout: '',
       stderr: '',
-      exitCode: processExitCode.spawn,
+      exitCode: String(processExitCode.spawn),
     });
 
     mockedNpxInPty.mockResolvedValue({
@@ -86,18 +87,7 @@ describe('ui:create:react', () => {
   };
 
   const doMockConfiguration = () => {
-    mockedConfig.mockImplementation(
-      () =>
-        ({
-          get: () =>
-            ({
-              environment: 'dev',
-              organization: 'my-org',
-              region: 'us-east-1',
-              analyticsEnabled: true,
-            } as Configuration),
-        } as Config)
-    );
+    mockedConfig.mockImplementation(configurationMock());
   };
 
   const doMockAuthenticatedClient = () => {
@@ -205,7 +195,7 @@ describe('ui:create:react', () => {
       expect(mockedNpxInPty).nthCalledWith(
         1,
         [
-          'create-react-app',
+          'create-react-app@1.0.0',
           'myapp',
           '--template',
           '@coveo/cra-template@1.0.0',