+++ packages/cli/src/lib/decorators/preconditions/npm.spec.ts
@@ -1,7 +1,6 @@
 jest.mock('../../utils/process');
 
 import {dedent} from 'ts-dedent';
-import {constants} from 'os';
 import {mocked} from 'ts-jest/utils';
 import {spawnProcessOutput} from '../../utils/process';
 import {getFakeCommand} from './testsUtils/utils';
@@ -33,7 +32,7 @@ describe('IsNpmVersionInRange', () => {
   describe('when npm is not installed', () => {
     beforeEach(() => {
       mockedSpawnProcessOutput.mockResolvedValue({
-        exitCode: constants.errno.ENOENT,
+        exitCode: 'ENOENT',
         stderr: '',
         stdout: '',
       });
@@ -56,7 +55,7 @@ describe('IsNpmVersionInRange', () => {
   describe('when an unknown error happens while getting the npm version', () => {
     beforeEach(() => {
       mockedSpawnProcessOutput.mockResolvedValue({
-        exitCode: 1,
+        exitCode: '1',
         stderr: 'some random error oh no',
         stdout: '',
       });
@@ -83,7 +82,7 @@ describe('IsNpmVersionInRange', () => {
   describe('when the installed version of npm is lower than the required one', () => {
     beforeEach(() => {
       mockedSpawnProcessOutput.mockResolvedValue({
-        exitCode: 0,
+        exitCode: '0',
         stderr: '',
         stdout: 'v0.9.0',
       });
@@ -109,7 +108,7 @@ describe('IsNpmVersionInRange', () => {
   describe('when the installed version of npm is above than the required one', () => {
     beforeEach(() => {
       mockedSpawnProcessOutput.mockResolvedValue({
-        exitCode: 0,
+        exitCode: '0',
         stderr: '',
         stdout: 'v1.1.0',
       });
@@ -128,7 +127,7 @@ describe('IsNpmVersionInRange', () => {
   describe('when the installed version of npm is the same as the required one', () => {
     beforeEach(() => {
       mockedSpawnProcessOutput.mockResolvedValue({
-        exitCode: 0,
+        exitCode: '0',
         stderr: '',
         stdout: 'v1.0.0',
       });