+++ packages/cli/src/lib/utils/misc.ts
@@ -8,8 +8,11 @@ export function getPackageVersion(packageName: string) {
     'package.json'
   );
   const pkg = require(pathToPackageJson);
-  const dep = pkg.dependencies[packageName] || pkg.devDependencies[packageName];
-  const defaultVersion = coerce(dep.toString())?.version;
+  const dep =
+    pkg.dependencies[packageName] ||
+    pkg.devDependencies[packageName] ||
+    pkg.peerDependencies[packageName];
+  const defaultVersion = coerce(dep?.toString())?.version;
 
   return defaultVersion;
 }